+++ bench/src/main/java/me/jinsui/shennong/bench/writer/HDFSWriter.java
@@ -49,7 +49,7 @@ public class HDFSWriter extends Writer {
                 "-r", "--rate"
             },
             description = "Write rate bytes/s across all file")
-        public int writeRate = 1000000;
+        public double writeRate = 1000000;
 
         @Parameter(
             names = {
@@ -218,7 +218,7 @@ public class HDFSWriter extends Writer {
             numBytesForThisThread);
 
         // Acquire 1 second worth of records to have a slower ramp-up
-        RateLimiter.create(flags.writeRate / flags.numThreads).acquire(flags.writeRate / flags.numThreads);
+        RateLimiter.create(flags.writeRate / flags.numThreads).acquire((int) (flags.writeRate / flags.numThreads));
 
         long totalWritten = 0L;
         long totalBytesWritten = 0L;
